Add UnixTimestampJsonConverter and refine data types
Added a new JsonConverter for Unix timestamps to be used throughout the project models. Also refined several data types in various locations, for example switching from `int` to `double` where necessary.

+++ NDJH.Hypixel.API/Models/SkyBlock/Bingo.cs
@@ -9,9 +9,9 @@ public record Goal(
     [property: JsonPropertyName("lore")] string Lore,
     [property: JsonPropertyName("tiers")] IReadOnlyList<int> Tiers,
     [property: JsonPropertyName("progress")]
-    int Progress,
+    long Progress,
     [property: JsonPropertyName("requiredAmount")]
-    int RequiredAmount);
+    long RequiredAmount);
 
 public record BingoResponse(
     [property: JsonPropertyName("id")] int Id,
